AuroraBackend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.2.0
amazon-transcribe==0.6.2
annotated-types==0.7.0
anyio==4.6.0
astunparse==1.6.3
attrs==23.1.0
audioplayer==0.6
awscrt==0.16.26
beautifulsoup4==4.12.3
blinker==1.6.3
boto3==1.34.154
botocore==1.34.154
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==43.0.1
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
filelock==3.16.1
Flask==3.0.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-speech==2.27.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.1
grpcio-status==1.66.1
gTTS==2.5.3
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.6
importlib-metadata==6.11.0
itsdangerous==2.1.2
jax==0.4.31
jaxlib==0.4.31
Jinja2==3.1.2
jiter==0.5.0
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kagglehub==0.3.4
keras==3.4.1
kiwisolver==1.4.5
labelImg==1.8.6
langchain==0.3.0
langchain-core==0.3.5
langchain-text-splitters==0.3.0
langsmith==0.1.125
libclang==18.1.1
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.1
mdurl==0.1.2
mediapipe==0.10.21
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.1.0
namex==0.0.8
networkx==3.4.2
nltk==3.8.1
numpy==1.26.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
packaging==23.2
pandas==2.1.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.1.0
playsound==1.3.0
proto-plus==1.24.0
protobuf==4.25.6
psutil==6.1.1
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==14.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.6.1
Pygments==2.17.2
pymongo==4.8.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
pyspark==3.5.4
python-dateutil==2.8.2
python-docx==1.1.2
python-pptx==1.0.2
pytz==2023.3.post1
PyYAML==6.0.2
pyzbar==0.1.9
referencing==0.31.1
regex==2023.10.3
requests==2.32.3
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
sentencepiece==0.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
soundfile==0.12.1
soupsieve==2.6
SQLAlchemy==2.0.35
streamlit==1.29.0
sympy==1.13.1
tenacity==8.2.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tokenizers==0.21.0
toml==0.10.2
toolz==0.12.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
ultralytics==8.3.58
ultralytics-thop==2.0.13
urllib3==2.1.0
utils==1.0.2
validators==0.22.0
watchdog==3.0.0
webrtcvad-wheels==2.0.11.post1
Werkzeug==3.0.0
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.11.1
zipp==3.17.0